refactor ContentModel#getOverwriteMetadata() to also support news, events + faq
