+++ calendar-bundle/src/Resources/contao/models/CalendarEventsModel.php
@@ -10,6 +10,7 @@
 
 namespace Contao;
 
+use Contao\CoreBundle\File\ModelMetadataTrait;
 use Contao\Model\Collection;
 
 /**
@@ -200,6 +201,8 @@ use Contao\Model\Collection;
  */
 class CalendarEventsModel extends Model
 {
+	use ModelMetadataTrait;
+
 	/**
 	 * Table name
 	 * @var string
