+++ faq-bundle/src/Resources/contao/models/FaqModel.php
@@ -10,6 +10,7 @@
 
 namespace Contao;
 
+use Contao\CoreBundle\File\ModelMetadataTrait;
 use Contao\Model\Collection;
 
 /**
@@ -112,6 +113,8 @@ use Contao\Model\Collection;
  */
 class FaqModel extends Model
 {
+	use ModelMetadataTrait;
+
 	/**
 	 * Table name
 	 * @var string
